Move PR puller logic into its own puller
